get_jaccard_index_for_three_databases Function

This function computes the Jaccard index, a measure of similarity for the genes listed in the results of three different databases (GO, KEGG, Reactome). It first extracts the relevant gene information from the provided object, then computes the Jaccard index between all pairs of modules across the three databases.

get_jaccard_index_for_three_databases(
  object,
  module_result_go,
  module_result_kegg,
  module_result_reactome
)

Arguments

object

An mass_dataset object containing gene information.

module_result_go

A data frame containing module results from GO database.

module_result_kegg

A data frame containing module results from KEGG database.

module_result_reactome

A data frame containing module results from Reactome database.

Value

A data frame with the Jaccard index values between all pairs of modules across the three databases.

Author

Xiaotao Shen shenxt1990@outlook.com