merge_modules Function

This function merges information from three databases (GO, KEGG, Reactome) and performs similarity analysis among pathways/modules across these databases. It filters, rearranges, and aggregates the data before computing the Jaccard Index for similarity. The function also creates network visualizations and saves the results to an Excel file.

merge_modules(
  object,
  module_result_go,
  module_result_kegg,
  module_result_reactome,
  sim.cutoff = 0.5,
  measure_method = c("jaccard"),
  path = "result"
)

Arguments

object

An mass_dataset object containing gene information.

module_result_go

A data frame containing module results from GO database.

module_result_kegg

A data frame containing module results from KEGG database.

module_result_reactome

A data frame containing module results from Reactome database.

sim.cutoff

A numeric value for similarity cutoff. Default is 0.5.

measure_method

A character vector specifying the method of similarity measurement. Default is "jaccard".

path

A character string specifying the directory to save results. Default is "result".

Value

Outputs an Excel file and a PDF containing the merged and analyzed data.

Author

Xiaotao Shen shenxt1990@outlook.com