Plot a Network of Modules, Pathways, and Molecules

This function generates a network plot based on the input data which includes module results, pathway enrichments, and marker information.

plot_network(
  module_result = module_result_all,
  result_with_module = result_with_module,
  enrichment_go = enrichment_go,
  enrichment_kegg = enrichment_kegg,
  enrichment_reactome = enrichment_reactome,
  marker_info = marker,
  including_molecule = TRUE,
  circle = FALSE,
  node_color = c(Function_module = "#F05C3BFF", Module = "#46732EFF", Pathway =
    "#197EC0FF", Molecule = "#3B4992FF"),
  text_size = 3,
  arrange_position = TRUE,
  position_ratio = 0.95
)

Arguments

module_result

A data frame of module results.

result_with_module

A data frame of results with module information.

enrichment_go

A data frame of GO enrichment results.

enrichment_kegg

A data frame of KEGG enrichment results.

enrichment_reactome

A data frame of Reactome enrichment results.

marker_info

A data frame of marker information.

including_molecule

Logical, whether to include molecules in the plot.

circle

Logical, whether to use a circular layout for the plot.

node_color

A vector of colors for different node types.

text_size

Numeric, size of the text in the plot.

arrange_position

Logical, whether to arrange position of nodes.

position_ratio

Numeric, ratio for arranging position of nodes.

Value

A ggplot object representing the network plot.

Examples

if (FALSE) {
plot_network(module_result,
result_with_module,
enrichment_go,
enrichment_kegg,
enrichment_reactome,
marker_info)
}