Perform pathway enrichment analysis

This function is a wrapper for enrich_pathway_internal. It dispatches the call to the appropriate method based on the class of the object argument. This functions is based on the clusterProfiler and ReactomePA package.

enrich_pathway(
  object,
  database = c("go", "kegg", "reactome"),
  path = "result",
  OrgDb,
  organism = "hsa",
  keyType = "ENTREZID",
  use_internal_data = FALSE,
  ont = "ALL",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe,
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  readable = FALSE,
  pool = FALSE,
  ...
)

# S3 method for mass_dataset
enrich_pathway(
  object,
  database = c("go", "kegg", "reactome"),
  path = "result",
  OrgDb,
  organism = "hsa",
  keyType = "ENTREZID",
  use_internal_data = FALSE,
  ont = "ALL",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe,
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  readable = FALSE,
  pool = FALSE,
  ...
)

# S3 method for data.frame
enrich_pathway(
  object,
  database = c("go", "kegg", "reactome"),
  path = "result",
  OrgDb,
  organism = "hsa",
  keyType = "ENTREZID",
  use_internal_data = FALSE,
  ont = "ALL",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe,
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  readable = FALSE,
  pool = FALSE,
  ...
)

Arguments

object

A data frame or mass_dataset object containing gene identifiers.

database

A vector of databases to use for enrichment (default: c("GO", "kegg", "reactome")).

path

The directory where results should be saved (default: "result").

OrgDb

The OrgDb object to use for GO enrichment.

organism

The organism code for KEGG and Reactome enrichment (default: "hsa").

keyType

The key type for gene identifiers (default: "ENTREZID").

use_internal_data

Logical, use internal data for KEGG enrichment (default: FALSE).

ont

The ontology to use for GO enrichment (default: "ALL").

pvalueCutoff

The p-value cutoff for enrichment (default: 0.05).

pAdjustMethod

The method for p-value adjustment (default: "BH").

universe

The background gene set.

qvalueCutoff

The q-value cutoff for enrichment (default: 0.2).

minGSSize

The minimum gene set size for enrichment (default: 10).

maxGSSize

The maximum gene set size for enrichment (default: 500).

readable

Logical, make the output readable (default: FALSE).

pool

Logical, pool genes for GO enrichment (default: FALSE).

...

Additional arguments passed to the underlying functions.