Identify Functional Modules Across Different Databases

This function identify functional modules from various databases, such as GO, KEGG, and Reactome. It calculates the similarity matrix between all the pathways and clusters them into functional modules.

merge_modules(
  variable_info,
  object,
  sim.cutoff = 0.5,
  measure_method = c("jaccard"),
  path = "result",
  save_to_local = FALSE
)

Arguments

variable_info

A data frame or tibble containing variable information.

object

An S4 object, expected to be processed by merge_pathways().

sim.cutoff

A numerical value for similarity cutoff, default is 0.5.

measure_method

A character vector indicating the method for measuring similarity, default is "jaccard".

path

A character string for the directory where to save results, default is "result".

save_to_local

Logical, if TRUE the results will be saved to local disk.

Note

Please ensure that object has been processed by merge_pathways() before using this function.

Author

Xiaotao Shen shenxt1990@outlook.com

Examples

if (FALSE) {
merge_modules(variable_info = my_var_info, object = my_object)
}